feat(sidebar context): exposing sidebar context hook
