+++ public/app/core/components/AppChrome/ExtensionSidebar/ExtensionSidebar.test.tsx
@@ -1,24 +1,21 @@
 import { render, screen } from '@testing-library/react';
 
 import { ExtensionInfo, PluginExtensionTypes } from '@grafana/data';
-import { config, usePluginComponents } from '@grafana/runtime';
+import {
+  config,
+  usePluginComponents,
+  useExtensionSidebarContext,
+  type ExtensionSidebarContextType,
+} from '@grafana/runtime';
 import { AddedComponentRegistryItem } from 'app/features/plugins/extensions/registry/AddedComponentsRegistry';
 import { createComponentWithMeta } from 'app/features/plugins/extensions/usePluginComponents';
 
 import { ExtensionSidebar } from './ExtensionSidebar';
-import {
-  ExtensionSidebarContextType,
-  getComponentIdFromComponentMeta,
-  useExtensionSidebarContext,
-} from './ExtensionSidebarProvider';
+import { getComponentIdFromComponentMeta } from './ExtensionSidebarProvider';
 
 jest.mock('@grafana/runtime', () => ({
   ...jest.requireActual('@grafana/runtime'),
   usePluginComponents: jest.fn(),
-}));
-
-jest.mock('./ExtensionSidebarProvider', () => ({
-  ...jest.requireActual('./ExtensionSidebarProvider'),
   useExtensionSidebarContext: jest.fn(),
 }));
