+++ public/app/core/components/AppChrome/ExtensionSidebar/ExtensionSidebar.tsx
@@ -2,12 +2,11 @@ import { css } from '@emotion/css';
 import { css as cssReact, Global } from '@emotion/react';
 
 import { GrafanaTheme2, PluginExtensionPoints } from '@grafana/data';
-import { usePluginComponents } from '@grafana/runtime';
+import { usePluginComponents, useExtensionSidebarContext } from '@grafana/runtime';
 import { useTheme2 } from '@grafana/ui';
 
-import { getComponentMetaFromComponentId, useExtensionSidebarContext } from './ExtensionSidebarProvider';
+import { getComponentMetaFromComponentId } from './ExtensionSidebarProvider';
 
-export const DEFAULT_EXTENSION_SIDEBAR_WIDTH = 300;
 export const MIN_EXTENSION_SIDEBAR_WIDTH = 100;
 export const MAX_EXTENSION_SIDEBAR_WIDTH = Math.floor(window.innerWidth * (2 / 3));
