+++ public/app/core/components/AppChrome/TopBar/SingleTopBarActions.tsx
@@ -2,11 +2,10 @@ import { css, cx } from '@emotion/css';
 
 import { GrafanaTheme2 } from '@grafana/data';
 import { Components } from '@grafana/e2e-selectors';
-import { ScopesContextValue } from '@grafana/runtime';
+import { ScopesContextValue, useExtensionSidebarContext } from '@grafana/runtime';
 import { Stack, useStyles2 } from '@grafana/ui';
 import { ScopesSelector } from 'app/features/scopes/selector/ScopesSelector';
 
-import { useExtensionSidebarContext } from '../ExtensionSidebar/ExtensionSidebarProvider';
 import { NavToolbarSeparator } from '../NavToolbar/NavToolbarSeparator';
 
 import { getChromeHeaderLevelHeight } from './useChromeHeaderHeight';
